+++ CHANGELOG.md
@@ -4,8 +4,34 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to Calendar Versioning (CalVer).
+<!-- ## Template
+
+### Added
+
+### Changed
+
+### Deprecated
+
+### Removed
+
+### Fixed
+
+### Security -->
 
 ## Unreleased
+
+### Added
+
+### Changed
+- Deleted bubbles style [PR #471](https://github.com/marmot-protocol/whitenoise/pull/471)
+
+### Deprecated
+
+### Removed
+
+### Fixed
+
+### Security
 
 ## [2026.3.5] - 2026-03-05